Migrated from json-file.
