+++ .eslintrc
@@ -1,0 +1,22 @@
+{
+  "rules": {
+    "indent": [2, 2],
+    "quotes": [2, "single"],
+    "linebreak-style": [2, "unix"],
+    "semi": [2, "always"],
+    "curly": [2, "multi-line"],
+    "handle-callback-err": [2, "^err"],
+    "strict": 0,
+    "eqeqeq": 0,
+    "eol-last": 0,
+    "dot-notation": 0,
+    "no-mixed-requires": 0,
+    "no-multi-spaces": 0,
+    "no-use-before-define": 0,
+    "no-underscore-dangle": 0
+  },
+  "env": {
+    "node": true,
+    "browser": true
+  }
+}
